Déplacement du parsing dans la classe xpl plutot que dans le message lui meme. Cela permet de séparer les parties émission et réception pour désactiver toute la partie réception/parsing si non nécessaire : #define ENABLE_PARSING dans xPL.h
Modification de plusieurs type


git-svn-id: http://xpl-arduino.googlecode.com/svn/trunk@9 6cc31b26-852d-9d07-f407-da9d70738961

+++ xPL_utils.cpp
@@ -31,7 +31,7 @@
 void clearStr (char* str)
 {
     int len = strlen(str);
-    for (int c = 0; c < len; c++)
+    for (byte c = 0; c < len; c++)
     {
         str[c] = 0;
     }
